+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/sticktothescript/gecooker/GECookerPlugin.java
@@ -1,16 +1,10 @@
-package net.runelite.client.plugins.sticktothescript.gecooker;
+package net.runelite.client.plugins.microbot.sticktothescript.gecooker;
 
 import com.google.inject.Provides;
 import lombok.extern.slf4j.Slf4j;
-import net.runelite.api.Client;
-import net.runelite.client.Notifier;
-import net.runelite.client.callback.ClientThread;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
-import net.runelite.client.plugins.microbot.Microbot;
-import net.runelite.client.plugins.microbot.util.mouse.VirtualMouse;
-import net.runelite.client.plugins.sticktothescript.gefiremaker.GEFiremakerOverlay;
 import net.runelite.client.ui.overlay.OverlayManager;
 
 import javax.inject.Inject;
