+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/sticktothescript/gefiremaker/GEFiremakerConfig.java
@@ -1,16 +1,15 @@
-package net.runelite.client.plugins.sticktothescript.gecooker;
+package net.runelite.client.plugins.microbot.sticktothescript.gefiremaker;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 import net.runelite.client.config.ConfigSection;
-import net.runelite.client.plugins.microbot.cooking.enums.CookingItem;
-import net.runelite.client.plugins.sticktothescript.common.enums.GEWorkLocation;
-import net.runelite.client.plugins.sticktothescript.common.enums.LogType;
+import net.runelite.client.plugins.microbot.sticktothescript.common.enums.GEWorkLocation;
+import net.runelite.client.plugins.microbot.sticktothescript.common.enums.LogType;
 
 
-@ConfigGroup("GECooker")
-public interface GECookerConfig extends Config {
+@ConfigGroup("GEFiremaker")
+public interface GEFiremakerConfig extends Config {
 
 
     @ConfigSection(
@@ -21,18 +20,18 @@ public interface GECookerConfig extends Config {
     String generalSection = "General";
 
     @ConfigSection(
-            name = "Cooking",
-            description = "Cooking Settings",
+            name = "Firemaking",
+            description = "Firemaking Settings",
             position = 1
     )
-    String cookingSection = "Cooking";
+    String firemakingSection = "Firemaking";
 
     @ConfigItem(
-            keyName = "Log Type",
+            keyName = "logType",
             name = "Log Type",
             description = "The type of logs to use to make the fire",
             position = 0,
-            section = cookingSection
+            section = firemakingSection
     )
     default LogType sLogType()
     {
@@ -40,23 +39,11 @@ public interface GECookerConfig extends Config {
     }
 
     @ConfigItem(
-            keyName = "Cook Item",
-            name = "Cook Item",
-            description = "What are we going to cook?",
-            position = 1,
-            section = cookingSection
-    )
-    default CookingItem sCookItem()
-    {
-        return CookingItem.RAW_SHRIMP;
-    }
-
-    @ConfigItem(
             keyName = "location",
             name = "Desired Fire Location",
             description = "The desired location to build a fire if a fire does not exist",
-            position = 2,
-            section = cookingSection
+            position = 1,
+            section = firemakingSection
     )
     default GEWorkLocation sLocation()
     {
@@ -64,11 +51,11 @@ public interface GECookerConfig extends Config {
     }
 
     @ConfigItem(
-            keyName = "Debug",
+            keyName = "debug",
             name = "Debug",
             description = "Enable debug information",
-            position = 3,
-            section = cookingSection
+            position = 2,
+            section = firemakingSection
     )
     default boolean sDebug()
     {
@@ -83,6 +70,6 @@ public interface GECookerConfig extends Config {
             section = generalSection
     )
     default String about() {
-        return "This plugin cooks the selected cooking item at the Grand Exchange using regular fires and campfires. If a fire does not already exist, it will create it in the desired location that you select. If you have any desired features, please contact me through Discord.";
+        return "This plugin adds logs to campfires at the Grand Exchange. If the fire does not already exist, it will create it in the desired location that you select. If you have any desired features, please contact me through Discord.";
     }
 }
