+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/sticktothescript/gefiremaker/GEFiremakerPlugin.java
@@ -1,15 +1,10 @@
-package net.runelite.client.plugins.sticktothescript.gefiremaker;
+package net.runelite.client.plugins.microbot.sticktothescript.gefiremaker;
 
 import com.google.inject.Provides;
 import lombok.extern.slf4j.Slf4j;
-import net.runelite.api.Client;
-import net.runelite.client.Notifier;
-import net.runelite.client.callback.ClientThread;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
-import net.runelite.client.plugins.microbot.Microbot;
-import net.runelite.client.plugins.microbot.util.mouse.VirtualMouse;
 import net.runelite.client.ui.overlay.OverlayManager;
 
 import javax.inject.Inject;
